requirements-dev.txt
-r requirements.txt
django-debug-toolbar==4.2.0
black==23.12.1
flake8==7.0.0
